Modificar Articulo, paginacion a Articulos.aspx
+ pequeñas correcciones, boton +, y estilo

+++ TPC_Comercio_Equipo26/TPC_Equipo26/Dominio/Imagen.cs
@@ -8,7 +8,7 @@ namespace TPC_Equipo26.Dominio
     public class Imagen
     {
         public long ID {get; set;}
-        public int IdArticulo { get; set; }
+        public long IdArticulo { get; set; }
         public string UrlImagen { get; set; }
         public bool Activo { get; set; }
     }
